make tests current
